SusyNtTools: refactor JVF requirement
Refactor isCentralLightJet and isSignalJet, moving the common lines
to a static function SusyNtTools::jetPassesJvfRequirement.
Verified that the outpout of `Susy2LepCF` is the same as before.
However, there is one detail that I should check with Steve:
`isCentralLightJet` used to have a fabs(jvf), which I don't understand.
In particular, if I add it to `jetPassesJvfRequirement`, there are a
handful of extra jets for which jvf==-1.0.